[Data Discovery] Add no results banner to samples view. (#2105)
Adds a no results banner to the samples view, just like the one that was added to projects and visualizations.

Tested it by choosing filters that lead to no results and confirming that the banner shows up. Confirmed that it does not show up when there is data in samples table.