Add new amr_counts endpoint for AMR Heatmap (#2311)
* Add /amr_heatmap.json endpoint (admin only)
Add a rails controller, amr_heatmap_controller.rb, and a rails template amr_hea tmap.html.erb. Modify routes.rb to add a GET endpoint for /amr_heatmap.json, which hits the index method of the AmrHeatmapController. This method fetches the amr counts for a given sample id and returns it as a JSON object. The actual amr counts are nested within the object, which also includes the sample name and id. The endpoint can only be accessed by admin accounts.

* Add Rspec tests for AmrHeatmapController and supporting Rspec factories
Created a new Rspec test for the AmrHeatmapController that tests if the controller is returning the correct json output for a given sample's amr_counts data and tests to make sure that the endpoint cannot be accessed by non-admin users.
Supporting Rspec factories were created to write this test, namely output_states.rb and amr_counts.rb. pipeline_runs.rb was modified to allow for creation of AmrCount classes and OutputState classes during the testing process.
In addition, the heatmap controller itself had a minor modification: user power :viewable_samples instead of power :samples.

* Fix GET request parameter for second AmrHeatmapController Rspec test

* Change AmrHeatmapController endpoint to accept multiple sample ids and implement PR comments
- remove CLI related code
- use current_power directly
- rename index endpoint to "amr_counts"
- accept multiple sample ids per request
- remove feature flag check
- remove unused Ruby view template
- remove unused code in Rspec test for unauthorized users
- removed use of Consul::Power in Rspec tests
- update comment on Rspec pipeline runs factory to encompass all transient params
- remove redundant comment
- remove unused :with_amr_count trait in Rspec pipeline_runs factory

* Change endpoint for AmrHeatmapController, expand test cases, add some error information
- change endpoint for amr_counts JSON to amr_heatmap/amr_counts.json
- add tests for cases where a sample id is invalid and a sample does not have amr_counts
- invalid sample ids return the same object but with an empty string for the sample name, an empty array for amr_counts, and a "sample not found" error message
- added error field in the returned json objects; if no error, should be an empty string
- changed the parameter for taking in sample ids from "id" to "sampleIds"
- fetch all samples in one query in the controller

* Add detailed comment describing what exactly amr_counts is

* Fix comment on AmrHeatmapController noting the route for amr_counts endpoint