Add upload_error to sample admin page (#2465)
Add upload_error to sample admin page# Description

This adds the new info to the admin page so it is obvious what is wrong. 

# Notes

I also changed the table style to highlight failures and de-emphasize common successes. 

# Tests

Open a sample that has a upload error
See text

![image](https://user-images.githubusercontent.com/28797/62336065-69c8e600-b483-11e9-91fc-2787c97bf311.png)


Open a sample that does not
See no text