Add test for report when taxon is missing from background model (#2766)
Add test for report when taxon is missing from background model# Description

Add a test for the new report service checking that the correct z-scores are returned when a taxon has a count type missing from the background model. 

If the count type is present in the sample, the z-score should be 100

If the count type is also missing in the sample, then the report will not contain that count type data for the taxon, and the 0 values should be filled in on the frontend instead.