Fetch unmapped reads unconditionally (#2840)
Fetch unmapped reads unconditionally# Description

The last PR to decrease fetching caused the unmapped reads to revert to the old faulty behavior. This PR reverts the code to the last known working version. 

I feel it is best to move on because these optimizations are difficult to test outside of staging, and the additional s3 fetches are few compared to the existing overfetching. See https://jira.czi.team/browse/IDSEQ-1924. 

# Tests

* check unmapped reads count in staging