[Report page] Update blue lightbulb and taxon row highlighting behavior (#2946)
[Report page] Update blue lightbulb and taxon row highlighting behavior# Description

These changes address the inconsistencies in how the old and new report pages highlight taxa brought up in IDSEQ-1974.

After some discussion with @katrinakalantar and @happyimadesignr, we decided to redefine what it meant for a row to be highlighted/to have the blue lightbulb to be more consistent. The top 3 species satisfying the thresholds will be highlighted and have the blue lightbulb, and the genera containing those species will be highlighted blue but **will not** be marked with the lightbulb.

This will hopefully better define why a particular row has been highlighted or marked (previously, it was possible for a species row to be highlighted but not its genus, and more than 3 rows could be marked with a lightbulb; the details behind this behavior wasn't well-documented) while still drawing the user's attention to taxa of interest.

This can mean that previously highlighted genus rows (without highlighted species) will no longer be highlighted, since only species will be checked against the thresholds. Additionally, genus rows will now **always** be highlighted if they contain species of interest, which previously was not necessarily the case.

![image](https://user-images.githubusercontent.com/53838890/72280185-9c5b3b00-35ec-11ea-99b0-de14aacd22f0.png)

![image](https://user-images.githubusercontent.com/53838890/72280190-9ebd9500-35ec-11ea-97bb-0772e572cd4d.png)


# Notes

The description and several images of the blue lightbulb on the Help Center will need to be updated to reflect these changes (specifically on [this page](https://chanzuckerberg.zendesk.com/hc/en-us/articles/360034790574-Single-Sample-Report-Table)).

# Tests

* Verified that the top 3 species marked on the legacy version of the report page for samples in project Medical Detectives have the blue lightbulb, and the genus rows are highlighted as expected.