Make sample open in new tab from heatmap (#2960)
Make sample open in new tab from heatmap# Description

![image](https://user-images.githubusercontent.com/28797/72466014-001c6a00-378d-11ea-8c14-7d3f54c4d4c3.png)

The current behavior was annoying because of the "are you sure you want to leave?" message. That is not triggered when opening in a new tab. Also, it seems like a user would typically want to compare the sample info with the heatmap and not go back and reload the entire heatmap. 

# Tests

* click on a square and see new tab without message