Remove dead code from pipeline monitor (#3101)
Remove dead code from pipeline monitor# Description

It looks like the original intent was restarting the job when the reason for batch termination was a EC2 instance being terminated (supposedly killed by an evicted spot instance, but could be something else). We already have auto-restarts in a different part of the code, so this is redundant.

The code is marked as "not being used" for more than 2 years, and adds unnecessary complexity. So I'm removing it.

# Tests

* None