Include exception type in autoscaling update error handling logs (#3137)
[Datadog] Include exception type in autoscaling update error logs# Description

Small fix related to https://github.com/chanzuckerberg/idseq-web/pull/3104.

The Datadog monitor for Python subprocess errors in the pipeline monitor explicitly looks for the error type `subprocess.CalledProcessError`, which wasn't getting printed out with the log tagged with `"[Datadog]"`. This change adds the error type so the monitor can be triggered as expected.