IDSEQ-2311: Remove Min Contigs Reads filter from report page (#3157)
IDSEQ-2311: Remove Min Contigs Reads filter from report page# Description

From the ticket: The min contig_r filter on the report page UI does not follow the same behavior as the other filters. It is also rarely used < 5 clicks / 1300 total filters applied in the past 3 months.

In addition, the original behavior of filtering on unique reads would be at odds with the new definition of reads as non-unique with the launch of the cdhitdup change. 

## Before

![image](https://user-images.githubusercontent.com/28797/76124089-ce669900-5fae-11ea-914c-3b08f8e6c55f.png)

## After 

![image](https://user-images.githubusercontent.com/28797/76124096-d1618980-5fae-11ea-9186-d8b1995ac1c5.png)

# Notes

* is it still possible to see the v1 report?
* is it worth to also clean up any related server side code?
* there are some screenshots here to change as well: https://chanzuckerberg.zendesk.com/hc/en-us/articles/360034790574-Single-Sample-Report-Table . Also note that the screen is already out of date because of the change from "size" to "reads". I'm not sure exactly how to change the screenshot. @happyimadesignr ?

# Tests

* load the report page v2, see no filter