[CH-62269] Temporarily persist settings from Heatmap -> Sample Report (#652)
# Description
Temporarily persist the Chosen background model, Threshold filters, Category filters, and Read Specificity filters from the Heatmap -> Sample Report when you click on a taxon cell.

The Sample Report temporarily applies these filters and does not carry the filters over when navigating to another sample report. When the user exits the sample report and goes back to it, the filters that were previously persisted temporarily from the Heatmap should not be there.

Also persist settings when the user opens up the sample details mode (by clicking on a sample name) and clicks "See report"

# Notes
`selectedOptions` from the `SampleView` & `SamplesHeatmapView` are not following the same structure and the metric names are different so creating an elegant solution to this was difficult. 

It would be great if we could consolidate the two into a synchronous format in the future

# Tests
1. Apply different filters in the heatmap then click the taxon cells.
2. The filters mentioned above should be the same in the SampleView && Heatmap.
3. Open a new tab and search for the sample that has the temporary filters on (this sample report should not have the temp filters)

Persisting two filters:
![temp filters](https://user-images.githubusercontent.com/51972068/105547865-d0634480-5cb3-11eb-94cc-b634e6f96714.gif)