[release-fix/CH-66865] Add projectId to globalContext when sample is fetched in sample report (#680)
# Description
According to feedback from @tfrcarvalho and @akrystosik, the projectId should be in the globalContext no matter how a user arrives at a sample. i.e. Link sharing, navigating via sample dropdown, copy & pasting URL...

# Notes
The `projectId` will be updated in the `globalContext` whenever a sample is now loaded so there is no need to pass it around in the URL anymore. For more info see conversations on this PR below

# Tests
* Generate `t` events, just like in #669 and ensure that the globalContext contains a valid projectId in the sample report and project samples view.
* If the user is not in a project or a sample report, then the `projectId` in the `globalContext` should be `null`. 
* Also: link sharing, navigating via sample dropdown, copy & pasting URL should also work and add the `projectId` in the `globalContext`

QA: Testing procedure is same as #669